requirements.txt
amqp==5.3.1
APScheduler==3.11.0
billiard==4.2.1
blinker==1.9.0
branca==0.8.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
folium==0.18.0
geojson==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
kombu==5.4.2
MarkupSafe==3.0.2
numpy==2.1.3
prompt_toolkit==3.0.48
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
requests==2.32.3
setuptools==75.6.0
six==1.16.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
xyzservices==2024.9.0
